style(condo): moved errors and helper functions from RegisterMeterReadingsService to another files
