+++ apps/condo/domains/meter/schema/RegisterMetersReadingsService.js
@@ -25,217 +25,12 @@ const {
     INVALID_METER_NUMBER,
     INVALID_DATE,
 } = require('@condo/domains/meter/constants/errors')
+const { ERRORS, READINGS_LIMIT } = require('@condo/domains/meter/constants/registerMetersReadingsService')
+const { isDateValid, toISO } = require('@condo/domains/meter/utils/date.utils')
+const { validateMeterValue, shouldUpdateMeter, meterReadingAsResult, normalizeMeterValue } = require('@condo/domains/meter/utils/meter.utils')
 const { Meter, MeterReading } = require('@condo/domains/meter/utils/serverSchema')
 
 dayjs.extend(customParseFormat)
-
-const ISO_DATE_FORMAT = 'YYYY-MM-DD'
-const EUROPEAN_DATE_FORMAT = 'DD.MM.YYYY'
-const DATE_PARSING_FORMATS = [
-    'YYYY-MM-DDTHH:mm:ss.SSS[Z]', // The result of dayjs().toISOString()
-    `${ISO_DATE_FORMAT} HH:mm:ss`, `${EUROPEAN_DATE_FORMAT} HH:mm:ss`, // Up to seconds
-    `${ISO_DATE_FORMAT} HH:mm`, `${EUROPEAN_DATE_FORMAT} HH:mm`, // Up to minutes
-    ISO_DATE_FORMAT, EUROPEAN_DATE_FORMAT, // No time
-    'YYYY-MM', 'MM-YYYY', 'YYYY.MM', 'MM.YYYY', // Some exotic cases
-]
-const READINGS_LIMIT = 500
-
-const ERRORS = {
-    TOO_MUCH_READINGS: {
-        code: BAD_USER_INPUT,
-        type: TOO_MUCH_READINGS,
-        message: 'Too much readings. {sentCount} sent, limit is {limit}.',
-        messageForUser: 'api.meter.registerMetersReadings.TOO_MUCH_READINGS',
-        messageInterpolation: {
-            limit: READINGS_LIMIT,
-            sentCount: '??',  // runtime value
-        },
-    },
-    ORGANIZATION_NOT_FOUND: {
-        code: BAD_USER_INPUT,
-        type: ORGANIZATION_NOT_FOUND,
-        message: 'Organization not found',
-        messageForUser: 'api.meter.registerMetersReadings.ORGANIZATION_NOT_FOUND',
-    },
-    PROPERTY_NOT_FOUND: {
-        code: BAD_USER_INPUT,
-        type: PROPERTY_NOT_FOUND,
-        message: 'Property not found',
-        messageForUser: 'api.meter.registerMetersReadings.PROPERTY_NOT_FOUND',
-    },
-    INVALID_METER_VALUES: {
-        code: BAD_USER_INPUT,
-        type: INVALID_METER_VALUES,
-        message: 'Invalid meter values',
-        messageForUser: 'api.meter.registerMetersReadings.INVALID_METER_VALUES',
-    },
-    MULTIPLE_METERS_FOUND: {
-        code: BAD_USER_INPUT,
-        type: MULTIPLE_METERS_FOUND,
-        message: 'Multiple meters found',
-        messageForUser: 'api.meter.registerMetersReadings.MULTIPLE_METERS_FOUND',
-    },
-    INVALID_ACCOUNT_NUMBER: {
-        code: BAD_USER_INPUT,
-        type: INVALID_ACCOUNT_NUMBER,
-        message: 'Invalid account number',
-        messageForUser: 'api.meter.registerMetersReadings.INVALID_ACCOUNT_NUMBER',
-    },
-    INVALID_METER_NUMBER: {
-        code: BAD_USER_INPUT,
-        type: INVALID_METER_NUMBER,
-        message: 'Invalid meter number',
-        messageForUser: 'api.meter.registerMetersReadings.INVALID_METER_NUMBER',
-    },
-    INVALID_DATE: {
-        code: BAD_USER_INPUT,
-        type: INVALID_DATE,
-        message: 'Invalid date',
-        messageForUser: 'api.meter.registerMetersReadings.INVALID_DATE',
-    },
-}
-
-const SYMBOLS_TO_CUT_FROM_DATES_REGEXP = /[^-_:.,( )/\d]/g
-
-/**
- * @param {string} dateStr
- * @return {undefined|string}
- */
-function sanitizeDateStr (dateStr) {
-    if (isEmpty(dateStr) || !isString(dateStr)) {
-        return undefined
-    }
-
-    const sanitizedDateStr = dateStr.replace(SYMBOLS_TO_CUT_FROM_DATES_REGEXP, '')
-    if (isEmpty(sanitizedDateStr)) {
-        return undefined
-    }
-
-    return sanitizedDateStr
-}
-
-function monthIsLoweOrEqualTwelve (dateStr) {
-    const [dateData] = dateStr.split(/T|\s/, 1)
-    const items = dateData.split(/[/\\\-.]/)
-
-    // YYYY MM || MM YYYY
-    if (items.length === 2) {
-        const [month] = items.sort((a, b) => a.length - b.length)
-        if (Number(month) > 12) {
-            return false
-        }
-    }
-
-    // YYYY MM DD || DD MM YYYY ...
-    if (items.length === 3) {
-        const month = items[1]
-        if (Number(month) > 12) {
-            return false
-        }
-    }
-
-    return true
-}
-
-/**
- * @param {string} dateStr
- * @return {undefined|string}
- */
-function toISO (dateStr) {
-    const sanitizedDateStr = sanitizeDateStr(dateStr)
-
-    if (isEmpty(sanitizedDateStr)) {
-        return undefined
-    }
-    return dayjs(sanitizedDateStr, DATE_PARSING_FORMATS).toISOString()
-}
-
-/**
- * @param {string} dateStr
- * @return {boolean}
- */
-function isDateValid (dateStr) {
-    const sanitizedStr = sanitizeDateStr(dateStr)
-    return dayjs(sanitizedStr, DATE_PARSING_FORMATS).isValid() && monthIsLoweOrEqualTwelve(sanitizedStr)
-}
-
-/**
- * @param value
- * @return {undefined|string|null}
- */
-function normalizeMeterValue (value) {
-    if (!(isString(value) || isNumber(value) || isUndefined(value))) return null
-    if (isUndefined(value)) return value
-
-    const transformedValue = String(value).trim().replaceAll(',', '.')
-    if (isEmpty(transformedValue)) return undefined
-    return String(Number(transformedValue))
-}
-
-/**
- * @param {string | null | undefined} value
- * @return {boolean}
- */
-function validateMeterValue (value) {
-    if (!isString(value) && !isUndefined(value)) return false
-    if (isUndefined(value)) return true
-
-    return !isEmpty(value) && !isNaN(Number(value)) && isFinite(Number(value)) && Number(value) >= 0
-}
-
-/**
- * @param {MeterReading} meterReading
- * @return {Object}
- */
-function meterReadingAsResult (meterReading) {
-    return {
-        id: meterReading.id,
-        meter: {
-            ...pick(meterReading.meter, ['id', 'unitType', 'unitName', 'accountNumber', 'number']),
-            property: pick(meterReading.meter.property, ['id', 'address', 'addressKey']),
-        },
-    }
-}
-
-function transformToPlainObject (input) {
-    let result = {}
-
-    for (let key in input) {
-        if (typeof input[key] === 'object' && input[key] !== null && input[key].id) {
-            result[key] = input[key].id
-        } else {
-            result[key] = input[key]
-        }
-    }
-
-    return result
-}
-
-/**
- * @param {Meter} meter
- * @param {RegisterMetersReadingsMeterMetaInput} changedFields
- * @return {boolean}
- */
-function shouldUpdateMeter (meter, changedFields) {
-    const fieldsToUpdate = [
-        'accountNumber',
-        'numberOfTariffs',
-        'place',
-        'verificationDate',
-        'nextVerificationDate',
-        'installationDate',
-        'commissioningDate',
-        'sealingDate',
-        'controlReadingsDate',
-    ]
-
-    return fieldsToUpdate.reduce((result, field) => {
-        if (result) {
-            return result
-        }
-        return !!get(changedFields, field) && get(changedFields, field) !== get(meter, field)
-    }, false)
-}
 
 const RegisterMetersReadingsService = new GQLCustomSchema('RegisterMetersReadingsService', {
     types: [
